✨FEAT: Añadir acciones de documentos a las versiones

+++ app/Filament/Actions/Documents/DeleteAction.php
@@ -2,6 +2,7 @@
 
 namespace App\Filament\Actions\Documents;
 
+use App\Models\File;
 use Filament\Actions\Action;
 use Filament\Actions\DeleteAction as FilamentDeleteAction;
 use Illuminate\Database\Eloquent\Model;
@@ -13,9 +14,15 @@ class DeleteAction
     {
         return FilamentDeleteAction::make()
             ->using(function (Model $record) {
-                $record->files()->each(function ($record) {
+                if ($record instanceof File) {
                     Storage::delete($record->path);
                     $record->delete();
+                    return;
+                }
+
+                $record->files()->each(function ($file) {
+                    Storage::delete($file->path);
+                    $file->delete();
                 });
 
                 $record->delete();
