+++ app/Filament/Actions/Documents/DeleteBulkAction.php
@@ -2,6 +2,7 @@
 
 namespace App\Filament\Actions\Documents;
 
+use App\Models\File;
 use Filament\Actions\Action;
 use Filament\Actions\DeleteBulkAction as FilamentDeleteBulkAction;
 use Illuminate\Support\Collection;
@@ -14,9 +15,15 @@ class DeleteBulkAction
         return FilamentDeleteBulkAction::make()
             ->using(function (Collection $records) {
                 $records->each(function ($record) {
-                    $record->files()->each(function ($record) {
+                    if ($record instanceof File) {
                         Storage::delete($record->path);
                         $record->delete();
+                        return;
+                    }
+
+                    $record->files()->each(function ($file) {
+                        Storage::delete($file->path);
+                        $file->delete();
                     });
 
                     $record->delete();
